fix: user 테이블에 social 속성 추가 및 테이블, 파일 이름 변경
